Added styles folder and moved css into folder
