+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 
-import "./Header.css";
+import "../styles/Header.css";
 
 export default function Header(props) {
